remove warning
git-svn-id: http://svn.cern.ch/guest/lhcb/Lbcom/trunk@26547 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: c65efa7c1b2f02ac834b81d69e106ba2d13ca687

+++ Lbcom/ST/STDAQ/src/RawBankToSTClusterAlg.cpp
@@ -1,4 +1,4 @@
-// $Id: RawBankToSTClusterAlg.cpp,v 1.7 2006-03-05 14:36:00 mneedham Exp $
+// $Id: RawBankToSTClusterAlg.cpp,v 1.8 2006-03-06 12:54:19 mneedham Exp $
 
 #include <algorithm>
 
@@ -164,7 +164,7 @@ StatusCode RawBankToSTClusterAlg::createCluster(const STClusterWord& aWord,
 
   STCluster::ADCVector adcs ; 
   for (unsigned int i = 1; i < adcValues.size() ; ++i){
-    adcs.push_back(std::make_pair(i-offset,adcValues[i].adc()));
+    adcs.push_back(std::make_pair(i-offset,(int)adcValues[i].adc()));
   } // iDigit
 
   // make cluster +set things
